flask-backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.1
Werkzeug==3.0.1
SQLAlchemy==2.0.27
bcrypt==4.1.2
PyJWT==2.8.0
